feat(core): chunked replay storage, realtime parity, presign hostname, fixes
- Replay: time-based S3 chunks (immutable gzip), chunk flush interval,
  detail API replay_chunk_urls + frontend stitch; legacy bundle fallback.
- Bun: idleTimeout, max body size, tracker/collect idle timeout exempt.
- S3: presign URLs via S3_PUBLIC_ENDPOINT for browser-reachable MinIO.
- Realtime: 30m pageview-scoped stats, timeline/breakdowns; recent activity
  within_minutes + activity→activities mapping.
- Heatmaps: normalize last_seen from Postgres (string vs Date) for /pages.
- Analytics SQL: numeric cast before round() where needed.
- Web: replay chunk fetch, realtime stat math, 128mb collect body hint.

Made-with: Cursor

+++ core/services/analytics/dimensions.ts
@@ -49,8 +49,10 @@ async function topDimensionAnalytics(
       CASE
         WHEN count(DISTINCT pv.session_id) = 0 THEN 0::float
         ELSE round(
-          sum(CASE WHEN spc.pvc = 1 THEN 1 ELSE 0 END)::float
-          / count(DISTINCT pv.session_id)::float * 100,
+          (
+            sum(CASE WHEN spc.pvc = 1 THEN 1 ELSE 0 END)::float
+            / count(DISTINCT pv.session_id)::float * 100
+          )::numeric,
           1
         )
       END AS bounce_rate
