+++ core/services/analytics/pages.ts
@@ -40,8 +40,11 @@ export async function getPagesAnalytics(
       CASE
         WHEN count(DISTINCT pv.session_id) = 0 THEN 0::float
         ELSE round(
-          sum(CASE WHEN spc.pvc = 1 THEN 1 ELSE 0 END)::float
-          / count(DISTINCT pv.session_id)::float * 100, 1
+          (
+            sum(CASE WHEN spc.pvc = 1 THEN 1 ELSE 0 END)::float
+            / count(DISTINCT pv.session_id)::float * 100
+          )::numeric,
+          1
         )
       END AS bounce_rate
     FROM pv
